+++ packages/otfjs-ui/src/components/icon-button/icon-button.tsx
@@ -1,0 +1,5 @@
+import styles from './icon-button.module.css'
+
+export function IconButton(props: React.HTMLAttributes<HTMLButtonElement>) {
+  return <button {...props} className={styles.root} />
+}
