+++ packages/otfjs-ui/src/components/icons/icon-back.tsx
@@ -1,0 +1,13 @@
+export function IconBack(props: React.SVGProps<SVGSVGElement>) {
+  return (
+    <svg {...props} width="24" height="24" viewBox="0 0 24 24" fill="none">
+      <path
+        stroke="var(--color-icon)"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth="1.5"
+        d="m15 18-6-6 6-6"
+      />
+    </svg>
+  )
+}
