+++ packages/otfjs-ui/src/utils/bytes.ts
@@ -1,0 +1,12 @@
+const SIZE_UNITS = ['B', 'KB', 'MB', 'GB']
+const SIZE_CONSTANT = 1024
+
+export function sizeToSTring(size: number) {
+  let unit = 0
+  while (size >= SIZE_CONSTANT && unit < SIZE_UNITS.length - 1) {
+    size /= SIZE_CONSTANT
+    ++unit
+  }
+
+  return `${size.toFixed(1)} ${SIZE_UNITS[unit]}`
+}
